+++ ruoyi-system/src/main/java/com/ruoyi/system/domain/dto/ProductSkuDTO.java
@@ -8,6 +8,7 @@ import lombok.Data;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Positive;
+import java.math.BigDecimal;
 
 @Data
 public class ProductSkuDTO {
@@ -21,7 +22,7 @@ public class ProductSkuDTO {
     @NotNull(message="价格不能为空")
     @Positive(message = "价格必须为正数")
     @ApiModelProperty(value = "价格")
-    private Long price;
+    private Double price;
 
     @ApiModelProperty(value = "价格单位")
     private String priceUnit;
